+++ lib/src/plugin/plugin.dart
@@ -8,4 +8,7 @@ abstract class BasePlugin {
 
   FutureOr<void> onBotStart(INyxx nyxx, Logger logger) async {}
   FutureOr<void> onBotStop(INyxx nyxx, Logger logger) async {}
+
+  FutureOr<void> onConnectionClose(INyxx nyxx, Logger logger, int closeCode, String? closeReason) async {}
+  FutureOr<void> onConnectionError(INyxx nyxx, Logger logger, String errorMessage) async {}
 }